zendesk_app/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
websockets==12.0
python-dotenv==1.0.0
